requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
djangorestframework==3.15.2
idna==3.10
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.1
urllib3==2.2.3
